test: seperate files for each method
